Move _withCanelability to the appropriate location.
