+++ Types/_source/LazyMixin.ts
@@ -76,27 +76,6 @@ export default abstract class LazyMixin {
 
         return result;
     }
-
-    /**
-     * Предоставляет возможность отменить Promise как это было реализовано в Deferred.
-     * @param main Основной вызов
-     * @param additional Дополнительный вызов
-     * @protected
-     */
-    protected _withCanelability<TResult>(
-        awaiter: Promise<TResult>,
-        callback: (result: TResult) => TResult
-    ): Promise<TResult> {
-        const result = awaiter.then(callback);
-
-        (result as IDeferred<TResult>).cancel = () => {
-            if ((awaiter as IDeferred<TResult>).cancel) {
-                (awaiter as IDeferred<TResult>).cancel();
-            }
-        };
-
-        return result;
-    }
 }
 
 Object.assign(LazyMixin.prototype, {
