+++ Types/_source/Remote.ts
@@ -23,7 +23,7 @@ import {
 import { RecordSet } from '../collection';
 import { create } from '../di';
 import { mixin, logger } from '../util';
-import { EntityMarker } from '../_declarations';
+import { EntityMarker, IDeferred } from '../_declarations';
 
 // tslint:disable-next-line:ban-comma-operator
 const global = (0, eval)('this');
@@ -286,7 +286,7 @@ export default abstract class Remote extends mixin<
             callResult,
             (data) => this._prepareCreateResult(data)
         );
-}
+    }
 
     read(key: EntityKey, meta?: object): Promise<Model> {
         const callResult = this._withAdditionalDependencies(
@@ -459,6 +459,27 @@ export default abstract class Remote extends mixin<
      */
     protected _prepareProviderArguments(args: object): object {
         return jsonize(args) as object;
+    }
+
+    /**
+     * Предоставляет возможность отменить Promise как это было реализовано в Deferred.
+     * @param main Основной вызов
+     * @param additional Дополнительный вызов
+     * @protected
+     */
+    protected _withCanelability<TResult>(
+        awaiter: Promise<TResult>,
+        callback: (result: TResult) => TResult
+    ): Promise<TResult> {
+        const result = awaiter.then(callback);
+
+        (result as IDeferred<TResult>).cancel = () => {
+            if ((awaiter as IDeferred<TResult>).cancel) {
+                (awaiter as IDeferred<TResult>).cancel();
+            }
+        };
+
+        return result;
     }
 
     protected _getValidKeyProperty(data: any): string {
